feat: json now fails if response is not in JSON format.
Sometimes the back end service can fail making it a 200 OK response with the error message as plain text.
`json` now will always expect a JSON back.
Since we might be wrongly using `json` to get plain text data this introduces a breaking change.
If we ever feel the need for a `text` method we will add it in a next version.